loading settings db, cache or config file.
we will load the settings from db then we will cache them so to avoid calling db again and again.

we will create a config file settings where we will add some default settings, which we will load if there are not settings in db or in cache to avoid the application to be crashed.